Remove oj gem from dependency
There are different customization needs when decoding JSON. For example,
in Rails, you would want to use Oj.optimize_rails(). And in other place,
you may just want the decoder to be as fast as possible.

Moreover, there are cases where we simply cannot use oj.

Therefore, in this gem, we decided to implement the to_json method with
the standard json library to avoid any debate